Basic Unit, Basic Action, Increased Map Scale
Created a basic unit class to store the various attributes of a unit.

Created a basic action class to store the various attributes of an
action.

Increased the scale of tile maps. Instead of using 32x32 tiles, I'll be
using 128x128 tiles for the map, with units being able to either be
128x128 textures or 192x192 textures for the larger units., though
having a larger texture will have no effect on gameplay. They'll just
take up more space on screen.